assignment_tracker/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
django-crontab==0.7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
